Extract tracing and branching methods into new helper class
